Create count_down_timer
I created a count-down timer that asks the user to input the amount of minutes and seconds they want their timer to count down from. Using an ActionListener, I was able to implement the 'start timer' button and I also used the JFrame to create the design for the application of the timer.