Create quiz
In this quiz file, I implemented a system that asks a user 4 questions that have A, B, C, and D answers. The quiz sets a timer for the user to respond in 10 seconds and automatically generates a wrong answer if they fail to respond in time. If they get a right answer, the answer turns green and if they get a wrong answer, it turns red. At the end of the quiz, the user gets to see how many they got right and wrong and get a percentage for the amount they got.