Create student_data
In this class, I engineered the main code that asks a user to enter input to find a students' information inside a database. The code asks the user for the students' name, grade level, courses, and the amount of payment and then generates all their data including the amount they have due left in an organized list.